refactor: MailForm, MailFormView 분리
